Analytics for population students show&tabs
Analytics now tracks students table show/hide button and tab changes.
