+++ services/oodikone2-frontend/src/common/hooks.js
@@ -1,7 +1,9 @@
-import { useState, useEffect, useRef } from 'react'
+import React, { useCallback, useState, useEffect, useRef } from 'react'
 import { chunk, isEqual } from 'lodash'
 import qs from 'query-string'
+
 import { SEARCH_HISTORY_VERSION } from '../constants'
+import TSA from './tsa'
 
 export const usePrevious = value => {
   const ref = useRef()
@@ -9,6 +11,24 @@ export const usePrevious = value => {
     ref.current = value
   })
   return ref.current
+}
+
+export const useTabChangeAnalytics = (category, action) => {
+  const previousTabIndex = React.useRef(0)
+
+  const handleTabChange = useCallback(
+    (e, data) => {
+      const { activeIndex, panes } = data
+
+      if (previousTabIndex.current !== activeIndex) {
+        TSA.Matomo.sendEvent(category, action, panes[activeIndex].menuItem)
+        previousTabIndex.current = activeIndex
+      }
+    },
+    [category, action, previousTabIndex]
+  )
+
+  return { handleTabChange }
 }
 
 export const useTabs = (id, initialTab, { location, replace }) => {
