+++ services/oodikone2-frontend/src/components/PopulationDetails/index.jsx
@@ -7,6 +7,7 @@ import { flattenDeep, intersection } from 'lodash'
 
 import selectors from '../../selectors/populationDetails'
 import { getTotalCreditsFromCourses } from '../../common'
+import { useTabChangeAnalytics } from '../../common/hooks'
 import PopulationFilters from '../PopulationFilters'
 import CreditAccumulationGraphHighCharts from '../CreditAccumulationGraphHighCharts'
 import CourseQuarters from '../CourseQuarters'
@@ -15,27 +16,6 @@ import PopulationCourses from '../PopulationCourses'
 import PopulationCreditGainTable from '../PopulationCreditGainTable'
 import InfoBox from '../InfoBox'
 import infoTooltips from '../../common/InfoToolTips'
-import TSA from '../../common/tsa'
-
-const sendAnalytics = (action, name, value) => TSA.Matomo.sendEvent('Population statistics', action, name, value)
-
-const useTabChangeAnalytics = action => {
-  const previousTabIndex = React.useRef(0)
-
-  const handleTabChange = useCallback(
-    (e, data) => {
-      const { activeIndex, panes } = data
-
-      if (previousTabIndex.current !== activeIndex) {
-        sendAnalytics(action, panes[activeIndex].menuItem)
-        previousTabIndex.current = activeIndex
-      }
-    },
-    [action, previousTabIndex]
-  )
-
-  return { handleTabChange }
-}
 
 const CourseStatisticsSegment = ({ samples, selectedStudents, translate }) => {
   const { CreditStatistics } = infoTooltips.PopulationStatistics
@@ -62,7 +42,7 @@ const CourseStatisticsSegment = ({ samples, selectedStudents, translate }) => {
     )
   }, [samples, selectedStudents, translate])
 
-  const { handleTabChange } = useTabChangeAnalytics('Change Credit statistics tab')
+  const { handleTabChange } = useTabChangeAnalytics('Population statistics', 'Change Credit statistics tab')
 
   return (
     <Segment>
